Fixed #5
