+++ src/CsvHelper.Excel.Specs/Parser/ParseUsingPathSpec.cs
@@ -1,0 +1,11 @@
+namespace CsvHelper.Excel.Specs.Parser
+{
+    public class ParseUsingPathSpec : ExcelParserSpec
+    {
+        public ParseUsingPathSpec() : base("parse_by_path")
+        {
+            using var parser = new ExcelParser(Path);
+            Run(parser);
+        }
+    }
+}
