+++ src/CsvHelper.Excel.Specs/Parser/ParseUsingPathSpecEmptyWithHeaders.cs
@@ -1,0 +1,11 @@
+namespace CsvHelper.Excel.Specs.Parser
+{
+    public class ParseUsingPathSpecEmptyWithHeaders : EmptySpecWithHeaders
+    {
+        public ParseUsingPathSpecEmptyWithHeaders() : base("parse_by_path_empty_with_headers")
+        {
+            using var parser = new ExcelParser(Path);
+            Run(parser);
+        }
+    }
+}
